% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axel_gdal.R
\name{axel}
\alias{axel}
\title{Download files via axel}
\usage{
axel(url, destfile, n, s, quiet)
}
\arguments{
\item{url}{A character string naming the URL of a resource to be downloaded.}

\item{destfile}{A character string with the name where the downloaded file will be saved.}

\item{n}{Specify maximum number of connections.}

\item{s}{Specify maximum speed (bytes per second).}

\item{quiet}{Leave stdout alone.}
}
\description{
Download files via axel
}
