% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axel_gdal.R
\name{axel_exist}
\alias{axel_exist}
\title{These functions provide information about the location 
of axel in your system.}
\usage{
axel_exist()
}
\description{
These functions provide information about the location 
of axel in your system.
}
