% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/to_be_removed_classifyData.R
\name{classifyData}
\alias{classifyData}
\title{Classification of remote sensing data}
\usage{
classifyData(bands, training_sites)
}
\description{
Classification of remote sensing data
}
\examples{
not run:
calcAtmosCorr(filepath = "Name_of_Landsat_Metadata_File")
}

