% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanPredictors.R
\docType{methods}
\name{cleanPredictors}
\alias{cleanPredictors}
\alias{cleanPredictors,GPM-method}
\title{Clean predictor variables regarding corelation and variance}
\usage{
\S4method{cleanPredictors}{GPM}(x, nzv = TRUE, highcor = TRUE,
  cutoff = 0.9, rmvna = TRUE)
}
\arguments{
\item{x}{An object of class gpm}

\item{nzv}{Remove (near) zero variability predictor variables}

\item{highcor}{Remove highly correlated predictor variables}

\item{cutoff}{Correlation cutoff value for highcor}

\item{rmvna}{Remove predictors with na values}
}
\value{
A layer within the gpm object with the information on the n 
individual resamplings.
}
\description{
Clean predictor variables regarding zero or near-zero variability or highly 
correlated variables.
}
\details{
The function is realized by using the caret::nearZeroVar and
caret::findCorrelation functions.
}
\examples{
\dontrun{
#Not run
}

}
\references{
The function uses functions from:
 Max Kuhn. Contributions from Jed Wing, Steve Weston, Andre Williams, 
 Chris Keefer, Allan Engelhardt, Tony Cooper, Zachary Mayer, Brenton Kenkel, 
 the R Core Team, Michael Benesty, Reynald Lescarbeau, Andrew Ziem, 
 Luca Scrucca, Yuan Tang and Can Candan. (2016). caret: Classification and 
 Regression Training. https://CRAN.R-project.org/package=caret
}
\seealso{
NONE
}
