% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compContTests.R
\name{compContTests}
\alias{compContTests}
\title{Compute contingency tests for all model responses}
\usage{
compContTests(models, mean = TRUE)
}
\arguments{
\item{models}{The trained model for each response variable and all 
resamplings resulting from \code{\link{trainModel}}}

\item{mean}{Compute the mean contingengy test results over all response 
variables or return an indivual result for each one of them}
}
\value{
A data frame containing the contingency test information.
}
\description{
Compute individual or averaged contingency tests for all predicted variables
}
\details{
NONE
}
\examples{
\dontrun{
#Not run
}

}
\references{
Contingency tests are taken from CAWCR (http://goo.gl/BEPWOI)
}
\seealso{
NONE
}
