% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compRegrTests.R
\name{compRegrTests}
\alias{compRegrTests}
\title{Compute regression tests over all model responses}
\usage{
compRegrTests(models, per_model = TRUE, per_selector = FALSE,
  sub_selectors = NULL, details = TRUE)
}
\arguments{
\item{NONE}{}
}
\description{
Compute regression tests over all model responses
}
\details{
NONE
}
\examples{
# Not run

}
\references{
NONE
}
\seealso{
NONE
}

