% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compVarImp.R
\name{compVarImp}
\alias{compVarImp}
\title{Compute variable importance over all model responses and resamplings}
\usage{
compVarImp(models, scale = FALSE)
}
\arguments{
\item{models}{The trained model for each response variable and all 
resamplings resulting from \code{\link{trainModel}}}

\item{scale}{Scale variable importance over all resamplings or use individual
values (TRUE/FALSE)}
}
\value{
A data frame containing the variable importance over for each
response variable and all resamplings.
}
\description{
Compute mean variable importance over all model responses and resamplings.
}
\details{
The variable importance is extracted from the model training dataset
based on functions supplied by the caret package.
}
\examples{
\dontrun{
#Not run
}

}
\references{
The function uses functions from:
 Max Kuhn. Contributions from Jed Wing, Steve Weston, Andre Williams, 
 Chris Keefer, Allan Engelhardt, Tony Cooper, Zachary Mayer, Brenton Kenkel, 
 the R Core Team, Michael Benesty, Reynald Lescarbeau, Andrew Ziem, 
 Luca Scrucca, Yuan Tang and Can Candan. (2016). caret: Classification and 
 Regression Training. https://CRAN.R-project.org/package=caret
}
\seealso{
NONE
}
