% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/conditionalSampleByVariable.R
\name{conditionalSampleByVariable}
\alias{conditionalSampleByVariable}
\title{Conditional sampling by specific variable}
\usage{
conditionalSampleByVariable(df, selector, grabs = 1, resample = 100)
}
\arguments{
\item{NONE}{}
}
\value{
NONE
}
\description{
This function creates n sets of conditional resamplings where the column
entries of a data frame are sampled based on a specific selector variable.
Within each sample, only m column entries for each unique value of the
selector variable are selected.
}
\details{
NONE
}
\examples{
# Not run
}
\references{
NONE
}
\seealso{
NONE
}

