% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createGPMMeta.R
\name{createGPMMeta}
\alias{createGPMMeta}
\title{Create a meta information dataset in gpm format}
\usage{
createGPMMeta(dataset, type = "input", selector, response, predictor, meta)
}
\arguments{
\item{dataset}{The name of the data frame variable used for initializing the 
gmp object}

\item{type}{The type of the dataset (e.g. input, output), defaults to "input"}

\item{selector}{An optional selector variable which can be used to control 
the sampling of subsets during model training}

\item{response}{The column ID of the response, i.e. dependent variable(s) in 
the dataset}

\item{predictor}{The column ID of the predictor, i.e. independent 
variable(s) in the dataset}

\item{meta}{The column ID of variables in the dataset which contain only meta
information not relevant for the model training}
}
\value{
A list in the appropriate format to be used as meta data layer in a
gpm object.
}
\description{
Create a meta dataset layer for a gpm object.
}
\details{
The column ID information is transformed to column names in order to 
ensure integrity even if columns are deleted in a later stage.
}
\examples{
\dontrun{
data(abies_alba)
createGPMMeta(dataset = abies_alba, type = "input",
              selector = 1, response = c(16:481), meta = c(2: 15))
}

}
\references{
NONE
}
\seealso{
\code{\link{AAAgpmClasses}} for the gpm class.
}
