% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axel_gdal.R
\name{gdaltranslate_exist}
\alias{gdaltranslate_exist}
\title{These functions provide information about the location 
of gdal_translate in your system.}
\usage{
gdaltranslate_exist()
}
\description{
These functions provide information about the location 
of gdal_translate in your system.
}
