% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gpm-package.R
\docType{package}
\name{gpm-package}
\alias{gpm}
\alias{gpm-package}
\title{Geospatial predictive modeling using parameterized and unparameterized models}
\description{
Geospatial predictive modeling using parameterized and unparameterized models
}
\details{
This package provides convenience functions for geospatial
predictive modeling of raster and saetllite data.
}
\author{
Tim Appelhans, Thomas Nauss, Hanna Meyer\cr
\cr
\emph{Maintainer:} Environmental Informatics \email{admin@environmentalinformatics-marburg.de}
}
\keyword{package}

