% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpm.R
\docType{methods}
\name{gpm}
\alias{gpm}
\alias{gpm,data.frame-method}
\title{Create a gpm object}
\usage{
\S4method{gpm}{data.frame}(x, meta, log, scale = FALSE, maxnew = 1,
  minnew = 0)
}
\arguments{
\item{x}{a data frame}

\item{meta}{meta information regarding the content of the columns, e.g. 
created using \code{\link{createGPMMeta}}}
}
\value{
gpm object
}
\description{
Method to create a gpm Object
}
\details{
A gpm object consists of three data sections:
(i) a data section which holds the actual dataset, (ii) a meta data grid 
which holds meta information for each of the columns in the data set and 
(iii) a list of log information which records the processing history of the 
data set.
}
\examples{
data(abies_alba)
meta <- createGPMMeta(dataset = abies_alba, 
                      selector = 1, response = c(16:481), meta = c(2: 15))
alb <- gpm(abies_alba, meta)

}
