% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpm_download.R
\name{gpm_download}
\alias{gpm_download}
\title{Download IMERG v05B Final, Early and Late Run}
\usage{
gpm_download(path, user, password, dates, product, band = 6,
  lonMin = -86, lonMax = -66, latMin = -19.25, latMax = 1.25,
  removeHDF5 = TRUE, quiet = TRUE, n = 2, s = 2.6)
}
\arguments{
\item{path}{path (character). The default corresponds to the working directory, getwd(). Missing values will be ignored.}

\item{user}{pmm user (character). See https://pmm.nasa.gov/data-access/downloads/gpm}

\item{password}{pmm password (character). See https://pmm.nasa.gov/data-access/downloads/gpm}

\item{dates}{sequences of days.}

\item{product}{IMERG product (character). Currently are supported: 'finalrun', 'late' and 'early'.}

\item{band}{band "HQobservationTime" = 1.
band "HQprecipSource" = 2.
band "HQprecipitation" = 3.
band "IRkalmanFilterWeight" = 4.
band "IRprecipitation" = 5.
band "precipitationCal" = 6.
band "precipitationUncal" = 7.
band "probabilityLiquidPrecipitation" = 8.
band "randomError" = 9.}

\item{lonMin}{Minimum latitude of bounding box}

\item{lonMax}{Maximum latitude of bounding box}

\item{latMin}{Minimum longitude of bounding box}

\item{latMax}{Maximum longitude of bounding box}

\item{removeHDF5}{Remove HD5 files after the conversion to GEOTIFF}

\item{quiet}{logical; suppress info of the download.}

\item{n}{Specify maximum number of connections (axel parameter)}

\item{s}{Specify maximum speed (megabytes per second)}
}
\description{
Download IMERG v05B Final, Early and Late Run
}
\details{
You need to have installed "Axel download accelerator
application" and GDAL. Obs: If you are using Windows,
the executables of "GDAL" and "AXEL" must be set as
system variables.

For install axel try the following:

Linux users: apt-get install axel

Windows axel: https://drive.google.com/uc?authuser=0&id=12T6R32-kAZPxbeCC3VdPx_KjqtxHWDt6

Use of the PPS FTP to download GPM and TRMM data is free, but you are required to first register your email address
http://registration.pps.eosdis.nasa.gov/registration/
You can found more about axel here: https://github.com/axel-download-accelerator/axel/issues.
}
\examples{
\dontrun{
gpm_download(path = '.',
            user = 'you_user@gmail.com',
            password = 'you_user@gmail.com',
            dates = c('2018-01-01'),
            product = 'finalrun')
}
}
\author{
Cesar Aybar
}
