% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axel_gdal.R
\name{gpm_getaxel}
\alias{gpm_getaxel}
\title{Donwnload axel 2.4 for windows user}
\usage{
gpm_getaxel(dirfile = getwd())
}
\arguments{
\item{dirfile}{A character string with the dirname where axel.zip well be saved.}
}
\description{
Donwnload axel 2.4 for windows user
}
