% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAAgpmClasses.R
\docType{class}
\name{GPMData-class}
\alias{GPMData-class}
\title{An S4 class to represent gpm data layers}
\description{
An S4 class to represent gpm data layers
}
\section{Slots}{

\describe{
\item{\code{data}}{a list object containing individual data objects}
}}

