% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAAgpmClasses.R
\docType{class}
\name{GPMLog-class}
\alias{GPMLog-class}
\title{An S4 class to represent gpm log data}
\description{
An S4 class to represent gpm log data
}
\section{Slots}{

\describe{
\item{\code{log}}{a list object containing information on individual processing steps}
}}

