% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAAgpmClasses.R
\docType{class}
\name{GPMMeta-class}
\alias{GPMMeta-class}
\title{An S4 class to represent gpm metadata information}
\description{
An S4 class to represent gpm metadata information
}
\section{Slots}{

\describe{
\item{\code{meta}}{a list object containing meta information}
}}

