% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAAgpmClasses.R
\docType{class}
\name{GPMModel-class}
\alias{GPMModel-class}
\title{An S4 class to represent gpm model layers}
\description{
An S4 class to represent gpm model layers
}
\section{Slots}{

\describe{
\item{\code{model}}{a list of model classes}
}}

