% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpStandardError.R
\name{helpStandardError}
\alias{helpStandardError}
\title{Calculate the standard error of the mean}
\usage{
helpStandardError(x)
}
\arguments{
\item{x}{a numeric vector}
}
\description{
This function calculates the standard error of the mean.
}
\examples{
a <- rnorm(1000, 20, 5)
mean(a)
se(a)
}

