% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/imputeData.R
\docType{methods}
\name{imputeData}
\alias{imputeData}
\alias{imputeData,GPM-method}
\alias{imputeData,data.frame-method}
\title{Impute missing data}
\usage{
\S4method{imputeData}{GPM}(x)

\S4method{imputeData}{data.frame}(x)
}
\arguments{
\item{NONE}{}
}
\value{
Imputed GPM object

Imputed data frame
}
\description{
Impute missing data




}
\details{
NONE
}
\examples{
# Not run
}
\references{
NONE
}
\seealso{
NONE
}

