% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lutInfo.R
\name{lutInfo}
\alias{lutInfo}
\title{Get or access internal LUT values used by various functions}
\usage{
lutInfo()
}
\arguments{
\item{None}{}
}
\value{
List containing several \code{data.frame} objects with LUT values.
}
\description{
Get internal look-up table (LUT) values from sysdata.rda which have been
compiled using data-raw/lut_data.R. Metadata is stored in \code{lut$meta}.
}
\details{
None
}
\examples{
None
}

