% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/meanOccupancy.R
\name{meanOccupancy}
\alias{meanOccupancy}
\title{Select features based on mean occupancy across unique locations}
\usage{
meanOccupancy(plotid, observations, resample = 100, thv = 20, seedid)
}
\arguments{
\item{NONE}{}
}
\value{
Columnnames of the selected features which occure at least on n
locations on average.
}
\description{
Compute occupancy of a feature x across y unique locations (but with
multiple samples per location) and return the feature names which occure at
least on n locations on average.
}
\details{
NONE
}
\examples{
# Not run
}
\references{
NONE
}
\seealso{
NONE
}

