% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/minimumOccurence.R
\name{minimumOccurence}
\alias{minimumOccurence}
\title{Select features based on minimum occurence across unique locations}
\usage{
minimumOccurence(plot_ids, observations, resample = 100, thv = 20)
}
\arguments{
\item{NONE}{}
}
\value{
Columnnames of the features occuring at least on n locations on
average.
}
\description{
Select features (e.g. species) based on their minimum occurence across unique
locations with multiple samples per location and return the feature names
which occure at least on n locations on average within multiple resamples
withour replacement.
}
\details{
NONE
}
\examples{
# Not run
}
\references{
NONE
}
\seealso{
NONE
}

