% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotClassPerformance.R
\name{plotClassPerformance}
\alias{plotClassPerformance}
\title{Plot performance of a classification model}
\usage{
plotClassPerformance(modstat, mean = FALSE, scores = c("Kappa", "ETS"))
}
\arguments{
\item{NONE}{}
}
\description{
Plot performance of a classification model
}
\details{
NONE
}
\examples{
# Not run

}
\references{
NONE
}
\seealso{
NONE
}
