% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVarImp.R
\name{plotVarImp}
\alias{plotVarImp}
\title{Plot mean variable importance over all model responses}
\usage{
plotVarImp(var_imp)
}
\arguments{
\item{NONE}{}
}
\description{
Plot mean variable importance over all model responses
}
\details{
NONE
}
\examples{
# Not run

}
\references{
NONE
}
\seealso{
NONE
}

