% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVarImpHeatmap.R
\name{plotVarImpHeatmap}
\alias{plotVarImpHeatmap}
\title{Plot mean variable importance over all model responses as heatmap}
\usage{
plotVarImpHeatmap(var_imp, xlab = "Variable", ylab = "Method",
  vis_range = "minmax")
}
\arguments{
\item{NONE}{}
}
\description{
Plot mean variable importance over all model responses as heatmap
}
\details{
NONE
}
\examples{
# Not run

}
\references{
NONE
}
\seealso{
NONE
}

