% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resamplingsByVariable.R
\docType{methods}
\name{resamplingsByVariable}
\alias{resamplingsByVariable}
\alias{resamplingsByVariable,GPM-method}
\alias{resamplingsByVariable,data.frame-method}
\title{Resampling by specific selector variable}
\usage{
\S4method{resamplingsByVariable}{GPM}(x, grabs = 1, resample = 100)

\S4method{resamplingsByVariable}{data.frame}(x, selector, grabs = 1,
  resample = 100)
}
\arguments{
\item{x}{An object of class gpm or data.frame}

\item{grabs}{The number m of samples per unique selector value}

\item{resample}{The number n of resamplings computed}

\item{selector}{The column name of the selector variable}
}
\value{
A layer within the gpm object with the information on the n 
individual resamplings.

A list of length n with the individual resamplings.
}
\description{
This function creates n sets of resamplings where the rows of a data frame 
are sampled based on a specific selector variable. The selector variable is
defined in the meta information layer of the gpm object. Within each sample, 
only m samples for each unique value of the selector variable are selected.
}
\details{
The resamplings do not contain the actual data values but the row 
numbers indicating which rows should be selected from the supplied data layer
during later processing steps.
}
\examples{
\dontrun{
#Not run
}

}
\references{
\code{\link{splitMultResp}} for splitting the samples into 
training and testing subsets.
}
\seealso{
NONE
}

