% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleBy.R
\name{sampleBy}
\alias{sampleBy}
\title{sample a Spatial*DataFrame by according to a factorial column}
\usage{
sampleBy(x, col, n = 1, seed = NULL, ...)
}
\arguments{
\item{x}{a Spatial*DataFrame}

\item{col}{the column used for the sampling}

\item{n}{the amount of samples to be taken from each level of col}

\item{seed}{an optional seed to be used for the 
random \code{\link{sample}}-ing.}

\item{...}{additional arguments passed to \code{\link{sample}}, 
e.g. \code{drop = ...}}
}
\value{
a list with entries \emph{sample_set} (the samples) 
and \emph{remainder} (the remainder).
}
\description{
This function samples a Spatial*DataFrame according to a specified
column so that all factor levels are sampled \code{n} times. If there
are less or equal entries in any levels these are sampled completely.
}
\examples{
# Not run
library(sp)

data(meuse)
coordinates(meuse) <- ~ x + y

sampleBy(meuse, col = "landuse", n = 1)

}
\seealso{
\code{\link{sample}} for details on the random sampling
}

