% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/splitByFrequency.R
\name{splitByFrequency}
\alias{splitByFrequency}
\title{splitByFrequency}
\usage{
splitByFrequency(csl, response, p = 0.75)
}
\arguments{
\item{NONE}{}
}
\value{
NONE
}
\description{
Compute
}
\details{
NONE
}
\examples{
# Not run
}
\references{
NONE
}
\seealso{
NONE
}

