% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/splitMultResp.R
\name{splitMultResp}
\alias{splitMultResp}
\title{Split dataset into testing and training samples for each response variable}
\usage{
splitMultResp(csl, response, p = 0.75)
}
\arguments{
\item{NONE}{}
}
\value{
NONE
}
\description{
Split a data set into testing and training samples separate for each response
variable while preserving the frequency distribution of the response variable.
}
\details{
NONE
}
\examples{
# Not run
}
\references{
NONE
}
\seealso{
NONE
}

