% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trainModel.R
\name{trainModel}
\alias{trainModel}
\title{Train machine learning model}
\usage{
trainModel(lodf, response_column, independent_columns, response_nbr = NULL,
  model_instc = NULL)
}
\arguments{
\item{NONE}{}
}
\value{
NONE
}
\description{
Train machine learning model.
}
\details{
NONE
}
\examples{
# Not run
}
\references{
NONE
}
\seealso{
NONE
}

