% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainModelrfe.R
\name{trainModelrfe}
\alias{trainModelrfe}
\title{Train machine learning model}
\usage{
trainModelrfe(resp, indp, n_var, mthd, seed_nbr, cv_nbr, metric,
  tune_length = NULL)
}
\arguments{
\item{NONE}{}
}
\value{
NONE
}
\description{
Train machine learning model.
}
\details{
NONE
}
\examples{
# Not run


}
\references{
NONE
}
\seealso{
NONE
}
