% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compContTests.R
\name{compContTests}
\alias{compContTests}
\title{Compute contingency tests over all model responses}
\usage{
compContTests(models)
}
\arguments{
\item{NONE}{}
}
\description{
Compute contingency tests over all model responses
}
\details{
NONE
}
\examples{
# Not run

}
\references{
Contingency tests are taken from CAWCR (http://goo.gl/BEPWOI)
}
\seealso{
NONE
}

