% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compVarImp.R
\name{compVarImp}
\alias{compVarImp}
\title{Compute variable importance over all model responses and resamplings}
\usage{
compVarImp(models)
}
\arguments{
\item{NONE}{}
}
\description{
Compute mean variable importance over all model responses and resamplings
}
\details{
NONE
}
\examples{
# Not run

}
\references{
NONE
}
\seealso{
NONE
}

