% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createGPMMeta.R
\name{createGPMMeta}
\alias{createGPMMeta}
\title{Create meta dataset}
\usage{
createGPMMeta(dataset, type = "input", selector, response, independent, meta)
}
\arguments{
\item{NONE}{}
}
\value{
NONE
}
\description{
Create a meta dataset layer for a gpm object.
}
\details{
NONE
}
\examples{
data(abies_alba)
createGPMMeta(dataset = abies_alba, type = "input",
              selector = 1, response = c(16:481), meta = c(2: 15))

}
\references{
NONE
}
\seealso{
NONE
}

