% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimumOccurence.R
\docType{methods}
\name{minimumOccurence}
\alias{minimumOccurence}
\alias{minimumOccurence,GPM-method}
\alias{minimumOccurence,data.frame-method}
\title{Select features based on minimum occurence across unique locations}
\usage{
\S4method{minimumOccurence}{GPM}(x, occurence = "yes", resample = 100,
  thv = 20)

\S4method{minimumOccurence}{data.frame}(x, selector, occurence = "yes",
  resample = 100, thv = 20)
}
\arguments{
\item{x}{An object of class gpm or data.frame}

\item{occurence}{Value indicating occurence (default "yes")}

\item{resample}{Number of resamples used to estimate the occurence 
(default 100)}

\item{thv}{Thresholdvalue which should be exceed in the occurence in order to
include the respective variable in the result list.}

\item{selector}{The column name of the selector variable which is used to
compute the minimum occurence distribution}
}
\value{
Columnnames of the features occuring at least on n locations on 
average as part of the meta data section of the gpm object

Columnnames of the features occuring at least on n locations on 
average.
}
\description{
Select features (e.g. species) based on their minimum occurence across unique
locations with multiple samples per location and return the feature names 
which occure at least on n locations on average within multiple resamples 
withour replacement.
}
\details{
The occurence is not counted but estimated based on samples which
are drawn from the dataset with one sample per selector ID.
}
\examples{
\dontrun{
#Not run
}

}
\references{
NONE
}
\seealso{
NONE
}

