% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotClassPerformance.R
\name{plotClassPerformance}
\alias{plotClassPerformance}
\title{Plot performance of a classification model}
\usage{
plotClassPerformance(model_stat, scores = c("KAPPA_MEAN", "TS_MEAN",
  "ETS_MEAN", "HK_MEAN"))
}
\arguments{
\item{NONE}{}
}
\description{
Plot performance of a classification model
}
\details{
NONE
}
\examples{
# Not run

}
\references{
NONE
}
\seealso{
NONE
}

