% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVarImp.R
\name{plotVarImp}
\alias{plotVarImp}
\title{Plot mean variable importance over all model responses}
\usage{
plotVarImp(var_imp)
}
\arguments{
\item{var_imp}{The resulting variable importance data rame from 
\code{\link{var_imp}}}
}
\value{
A list of visualizations with one visualization per response variable.
}
\description{
Plot mean variable importance over all model responses as horizontal bar plot.
}
\details{
NONE
}
\examples{
\dontrun{
#Not run
}

}
\references{
The function uses functions from:
 Max Kuhn. Contributions from Jed Wing, Steve Weston, Andre Williams, 
 Chris Keefer, Allan Engelhardt, Tony Cooper, Zachary Mayer, Brenton Kenkel, 
 the R Core Team, Michael Benesty, Reynald Lescarbeau, Andrew Ziem, 
 Luca Scrucca, Yuan Tang and Can Candan. (2016). caret: Classification and 
 Regression Training. https://CRAN.R-project.org/package=caret
}
\seealso{
NONE
}

