% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVarImpHeatmap.R
\name{plotVarImpHeatmap}
\alias{plotVarImpHeatmap}
\title{Plot mean variable importance over all model responses as heatmap}
\usage{
plotVarImpHeatmap(var_imp, xlab = "Variable", ylab = "Method",
  vis_range = "minmax")
}
\arguments{
\item{var_imp}{The resulting variable importance data rame from 
\code{\link{var_imp}}}

\item{xlab}{The x axis lable of the plot}

\item{ylab}{The y axis lable of the plot}

\item{vis_range}{Either "minmax" for a linear stretch over all values or a
vector of length two with the numeric minimum and maximum value.}
}
\value{
A list of heatmap visualizations with one visualization per response 
variable.
}
\description{
Plot mean variable importance over all model responses as heatmap
}
\details{
NONE
}
\examples{
\dontrun{
#Not run
}

}
\references{
NONE
}
\seealso{
NONE
}
