% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/resamplingsByVariable.R
\docType{methods}
\name{resamplingsByVariable}
\alias{resamplingsByVariable}
\alias{resamplingsByVariable,GPM-method}
\alias{resamplingsByVariable,data.frame-method}
\title{Resampling by specific variable}
\usage{
\S4method{resamplingsByVariable}{GPM}(x, grabs = 1, resample = 100)

\S4method{resamplingsByVariable}{data.frame}(x, selector, grabs = 1,
  resample = 100)
}
\arguments{
\item{NONE}{}
}
\value{
NONE

List with n resampled sets of row numbers from which the column
entries should be taken.

List with n resampled sets of row numbers from which the column
entries should be taken.
}
\description{
This function creates n sets of resamplings where the column
entries of a data frame are sampled based on a specific selector variable.
Within each sample, only m column entries for each unique value of the
selector variable are selected.




}
\details{
NONE
}
\examples{
# Not run
}
\references{
NONE
}
\seealso{
NONE
}

