% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/splitMultResp.R
\docType{methods}
\name{splitMultResp}
\alias{splitMultResp}
\alias{splitMultResp,GPM-method}
\alias{splitMultResp,data.frame-method}
\title{Split dataset into testing and training samples for each response variable}
\usage{
\S4method{splitMultResp}{GPM}(x, grabs = 1, resample = 100)

\S4method{splitMultResp}{data.frame}(x, response, resamples, p = 0.75)
}
\arguments{
\item{NONE}{}
}
\value{
NONE

List with n resampled sets of row numbers from which the column
entries should be taken.

List with n resampled sets of row numbers from which the column
entries should be taken.
}
\description{
Split a data set into testing and training samples separate for each response
variable while preserving the frequency distribution of the response variable.




}
\details{
NONE
}
\examples{
# Not run
}
\references{
NONE
}
\seealso{
NONE
}

