% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainModel.R
\docType{methods}
\name{trainModel}
\alias{trainModel}
\alias{trainModel,GPM-method}
\alias{trainModel,data.frame-method}
\title{Train machine learning model}
\usage{
\S4method{trainModel}{GPM}(x, grabs = 1, resample = 100)

\S4method{trainModel}{data.frame}(x, response, independent, resamples,
  mode = c("rfe", "ffs"), n_var = NULL, response_nbr = NULL,
  resample_nbr = NULL, mthd = "rf", seed_nbr = 11, cv_nbr = 2)
}
\arguments{
\item{NONE}{}
}
\value{
NONE

Trained model for each sample.

Trained model for each sample.
}
\description{
Train machine learning model.
}
\details{
NONE
}
\examples{
# Not run

}
\references{
NONE
}
\seealso{
NONE
}

